# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "GJG API",
    "description": "API documentation of GJG",
    "version": "v1"
  },
  "host": "localhost:8000",
  "schemes": [
    "http"
  ],
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Please write \"Bearer\" before the key"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
  "paths": {
...
    "/leaderboard/{country_iso_code}": {
      "get": {
        "operationId": "country_iso_code",
        "description": "",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": false,
            "type": "number",
            "default": 10
          },
          {
            "name": "skip",
            "in": "query",
            "description": "",
            "required": false,
            "type": "number",
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/LeaderBoard"
              }
            }
          }
        },
        "tags": [
          "Leader Board"
        ]
      },
      "parameters": [
        {
          "name": "country_iso_code",
          "in": "path",
          "description": "A country code.",
          "required": true,
          "type": "string"
        }
      ]
    },
...
  },
  "definitions": {
    "LeaderBoard": {
      "type": "object",
      "properties": {
        "rank": {
          "title": "Rank",
          "type": "integer",
          "readOnly": true,
          "default": 1
        },
        "points": {
          "title": "Points",
          "type": "integer",
          "readOnly": true,
          "default": 98565
        },
        "display_name": {
          "title": "Display name",
          "type": "string",
          "readOnly": true,
          "default": "gjg_user"
        },
        "country": {
          "title": "Countryy",
          "type": "string",
          "readOnly": true,
          "default": "tr"
        }
      }
    },
...
  }
}